Move cache and excpetions to api module
